- added global exception handling  - added pre-auth filters on the service layer  - added "manage fo" page to unclutter view fo (seperation of concern)  - streamlined the services along "manage fo"  - implemented all "manage fo" security function verification at the service layer
