+++ src/main/resources/templates/exception/forbiden-by-role.html
@@ -1,0 +1,18 @@
+<!DOCTYPE HTML>
+<html xmlns:th="http://www.thymeleaf.org">
+<head th:replace="~{fragments/general.html :: common_header(~{::title},~{},~{})}">
+<title>Data store - Home</title>
+</head>
+<body>
+	<div th:insert="~{fragments/general.html :: header_nav('none')}" />
+	<div class="container">
+		<h2>ERROR</h2>
+		<div class="alert alert-danger" role="alert">
+		  <span class="glyphicon glyphicon-exclamation-sign" aria-hidden="true"></span>
+		  <span class="sr-only">Error:</span>
+			  Your account does not have the proper role to access this page. 
+		</div>
+		<p>Please contact support central for any questions.</p>
+	</div>
+   </body>
+</html>
