+++ src/main/resources/templates/manage/addGrantingCapabilities.html
